+++ fission-core/library/Fission/Web/Auth/Token/JWT/RawContent.hs
@@ -1,13 +1,17 @@
 module Fission.Web.Auth.Token.JWT.RawContent
   ( contentOf
   , module Fission.Web.Auth.Token.JWT.RawContent.Types
+  , module Fission.Web.Auth.Token.JWT.RawContent.Class
   ) where
 
 import qualified RIO.Text                                    as Text
 
 import           Fission.Prelude
+
+import           Fission.Web.Auth.Token.JWT.RawContent.Class
 import           Fission.Web.Auth.Token.JWT.RawContent.Types
 
 -- | Smart constructor for 'RawContent'
 contentOf :: Text -> RawContent
+-- FIXME switch to HttpApiData, no?
 contentOf = RawContent . Text.dropEnd 1 . Text.dropWhileEnd (not . (== '.'))
